Cleanup code and fix some types
